Final version
This set of exercises finished. Final version of this file. 
Data was loaded, 
no cleaning was necessary, 
data was split to training and test data, 
then by using K-nearest neighbor a model was created, 
then model was tested, 
model was saved, 
saved model was tested to work,
finally 2 graphs were created to visualize the data.